Move DeleteOpenChannel into the OperatorUI dir
